Update portfolio.html
In h1, I added a shadow to the Projects header.  Makes it 'pop' out a little and stand out.
Added a :hover for the wordpress and discord buttons.  This makes it so when they are hovered on, they have a discord purple and a FreeBSD red backgrounds.
Added a fade between projects so when the button is pressed, one will fade out the the other will fade in smoothly.  I did this to my webpage with the images and it looks really good on yours!
Swapped the <p> and <img> classes so photo description is now before the image.  For me, it is just easier to see the description first so I know what I am looking at rather then look at it and figure out myself.
Made the photo description text 5px larger so it is a different size from the paragraphs.
Made the h2 text 2px larger to add some size variety.
Changed the background color to a more blueish color.  I just felt like there needs to be a little bit more color throughout.  This is mostly just a personal preference.  Not positive about the color I choose.  There might be a better fitting color to represent both BSD and Discord together better.